+++ packages/executors/src/index.ts
@@ -13,4 +13,5 @@ export * from "./retry.js";
 export * from "./seekai.js";
 export * from "./sse.js";
 export * from "./tabitoken.js";
+export * from "./tokenrouter.js";
 export * from "./search.js";